+++ wetterdienst/core/timeseries/request.py
@@ -193,6 +193,8 @@ class TimeseriesRequest(Core):
     #   - homogeneous parameters such as temperature_air_200
     interpolatable_parameters = [
         Parameter.TEMPERATURE_AIR_MEAN_200.name,
+        Parameter.TEMPERATURE_AIR_MAX_200.name,
+        Parameter.TEMPERATURE_AIR_MIN_200.name,
         Parameter.WIND_SPEED.name,
         Parameter.PRECIPITATION_HEIGHT.name,
     ]